Refactor into new service
